Add the Pages class and massively refactor PageResult
The Pages class is the next level up in the PaginatedResult
abstraction hierarchy above PageResult. It provides:
* a single constructor that takes a (copied) closure that retrieves
  successive pages from the API service
* iterator-returning begin() and end() methods.

The refactor for PageResult includes turning it into a nested class in
PageResult and removing all ElementAccessor members and parameters:
it is intended to be a lightweight, stateless, uniform wrapper, and is
now used purely as a default-constructed throwaway.